Move MoveFileByImportDate onto UtilActivities so it gets registered
The activity was a plain function in the workflow package, registered
nowhere, so enabling SortFilesByImportedDate would fail at runtime with
a missing activity. As a UtilActivities method it is registered on the
worker and debug queues like the other util activities. Also check the
generateRenameParams error before using its result.
